+++ documentation/examples/auto_gen_doc.rb
@@ -50,7 +50,7 @@ module AutoGenDoc
           # default_response 'default response', :json
           model = Object.const_get(action_path.split('#').first.split('/').last[0..-2].camelize) rescue nil
           type = action.in?(%w[ index show ]) ? Array[load_schema(model)] : String
-          response '200', 'success', :json, type: {
+          response '200', 'success', :json, data: {
               code: { type: Integer, dft: 200 },
                msg: { type: String,  dft: 'success' },
               data: { type: type }