+++ lib/oas_objs/request_body_obj.rb
@@ -11,10 +11,10 @@ module OpenApi
       attr_accessor :processed, :media_types
       def initialize(required, desc)
         self.media_types = [ ]
-        self.processed   = { required: required.to_s.match?(/req/), description: desc }
+        self.processed   = { required: required.match?('req'), description: desc }
       end
 
-      def add(media_type, hash)
+      def add_or_fusion(media_type, hash)
         media_types << MediaTypeObj.new(media_type, hash)
         self
       end
@@ -22,22 +22,6 @@ module OpenApi
       def process
         assign(media_types.map(&:process).reduce({ }, &fusion)).to_processed 'content'
         processed
-      end
-
-      def fusion
-        proc { |a, b| a.merge!(b, &_fusion) }
-      end
-
-      def _fusion
-        proc do |_common_key, x, y|
-          if x.is_a?(Hash) && y.is_a?(Hash)
-            x.merge(y, &_fusion)
-          elsif x.is_a?(Array) && y.is_a?(Array)
-            x.concat(y)
-          else
-            y
-          end
-        end
       end
     end
   end